#65 - fix README content/formatting
Some of the Markdown formatting didn't work in my viewer app, although it doesn't pose a problem for GitHub.  I think the app I use follows the Markdown "standards" more closely, but GH is lenient.

I also added a change that I'd mention to @ssciolla to clarify which directory "home" refers to.  (It's the user's home, not the project home.)