#65 - change MiVideo tables
MySQL likes to use TIMESTAMP type columns for special purposes.  In order to use that type in these tables, extra settings needed to be set to either provide default values for those columns or let them be null.  With DATETIME, it's easier to specify they should not be null and leave it at that.