#65 - Docker config improvements
Making it easier to work with the application locally in Docker with docker-compose.  Instead of copying the code to /app in Dockerfile, let docker-compose mount the project directory directly at /app.  Then, whenever the code is changed, it will be immediately available to the VM without a rebuild.  My addition to README.md documents a helpful development cycle.