#65 - more Docker config improvements
Reverting earlier Dockerfile change to let it use the command to `COPY` the project source code to `/app`.  That ensures deployments not using docker-compose will continue to work.  Docker seems to consider the `COPY` step as complete because the directory already contains the source code.  Also updated the development cycle documentation to reflect this change.