#92 - remove unnecessary Kaltura user ID
While experimenting with HJSON schema validation of the email address to be used in the Kaltura user ID parameter, I found the program continued to work even with random strings.  It turns out that Kaltura doesn't need the user ID.  That's not what the examples in their online documentation show, though.