#92 - move last time handling to function
Inspired by @ssciolla's comment in (https://github.com/tl-its-umich-edu/course-inventory/pull/116#discussion_r423125801).

I kind of felt the same, but the logging that I wanted to go along with the getting of the last time didn't feel like it belonged in the function.  However, since Sam also thought the handling should go into the function and I already have a little logging in other functions, I decided to go ahead and make the change.

My hesitation to add logging is that this function feels like an API function and I think logging doesn't belong in that kind of function.  🤷‍♀️