#142 - satisfy Pylint (Codacy)
I was not initially in favor of checking this type, because `pandas.Timestamp` is a subclass of `datetime.datetime`, but it seemed to be the right thing to do to prevent SQLAlchemy problems.  However, in looking up the "unidiomatic-typecheck" issue this line raised in Pylint, I learned the more pythonic technique is to embrace duck typing.  Maybe this type checking change should be undone, but I'll leave it in place for now as a teaching opportunity.